Add renderRecipe method to EmiRenderHelper
